backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tweepy==4.14.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
fasttext-langdetect==1.0.5
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
nltk==3.8.1
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.9
transformers==4.37.2
torch==2.2.0
aiohttp==3.9.3
asyncio==3.4.3
